adding artDeco with motd and message owl

+++ cli.py
@@ -1,9 +1,35 @@
 #!/usr/bin/env python
+# coding=latin-1
 
-import cmd, getpass, json, os, pprint, re, sys, time
+import cmd, getpass, json, os, pprint, re, sys, time, random
 import requests, execjs
 from lxml import html
 from bs4 import BeautifulSoup
+
+class ArtDeco():
+    bunnies = [
+        """  (\(\\\n (='.')\no(__")")""",
+        """(\__/)\n(='.'=)\n(")_(")""",
+        """ (\_/)\n=(^.^)=\n(")_(")""",
+        """(\ /)\n( . .)\nc(")(")""",
+        """(\__/)\n(>'.'<)\n(")_(")""",
+        """::: (\_(\\\n *: (=' :') :*\n•.. (,(”)(”)¤°.¸¸.•´¯`»""",
+        """   __//   \n  /.__.\  oops\n  \ \/ /  \n__/    \  \n\-      ) \n \_____/  \n___|_|____\n   " " """,
+        """|￣￣￣￣￣| \n|  LITTLE  |\n|  BUNNY   |\n|  LOVES   |\n|   YOU    |\n|    -xt2  |\n|＿＿＿＿＿| \n(\_/)  || \n(•ㅅ•) || \n/ 　 づ """,
+        """...........(\_/)\n...........( '_')........................ ☻...HELP!!!!\n...../++++++++++++++\======¦¦¦D -------- /▇ \\ \n/====================\..................  ||\n\_@___@___@___@___@__/"""
+    ]
+    def __init__( self ):
+        self.bunnyOTD = random.choice(self.bunnies)
+
+    def showMOTD( self ):
+        print self.bunnyOTD
+
+    def showMessage(self, msg):
+        self.__owl(msg)
+
+    def __owl( self, msg ):
+        print """,___,\n[O.o]  %s\n/)__)\n-"--"-\n""" % msg
+
 
 class Problem( object ):
     def __init__( self, pid, slug, level, tags=[], status=None, desc='', code='', test='' ):
@@ -338,7 +364,7 @@ class OJMixin( object ):
 
         return result
 
-class CodeShell( cmd.Cmd, OJMixin ):
+class CodeShell( cmd.Cmd, OJMixin, ArtDeco ):
     tags, problems, cheatsheet = {}, {}, {}
     tag = pid = sid = None
     test = '/tmp/test.dat'
@@ -502,11 +528,19 @@ class CodeShell( cmd.Cmd, OJMixin ):
             if not ( p.desc and p.code ):
                 p.desc, p.code, p.test = self.get_problem( p.slug )
             code = self.get_latest_solution( p )
+            if os.path.isfile( self.pad ):
+                self.showMessage("ARE YOU SURE?? (y/N)")
+                try:
+                    if raw_input().lower() not in [ 'y', 'yes']:
+                        print 'pull request cancelled.'
+                        return
+                except EOFError:
+                    print 'to exit, press ctrl+d again.'
+                    return
             with open( self.pad, 'w' ) as f:
                 f.write( code )
             with open( self.test, 'w' ) as f:
                 f.write( p.test )
-
         print self.pad
 
     def do_check( self, unused ):
